+++ packages/integrations/cloudflare/src/prerenderer.ts
@@ -61,6 +61,7 @@ export function createCloudflarePrerenderer({
 					outDir: fileURLToPath(serverDir),
 				},
 				root: fileURLToPath(root),
+				logLevel: 'error',
 				preview: {
 					host: 'localhost',
 					port: 0, // Let the OS pick a free port